requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
sqlalchemy
google-cloud-alloydb-connector
pg8000
